[Refactor] Refactor Create Protected Instances
