+++ .gitignore
@@ -2,3 +2,5 @@
 .idea
 .DS_Store
 .env
+*.env
+.vscode
